pyproject.toml
[build-system]
requires = ["uv-build>=0.6.6,<0.7"]
build-backend = "uv_build"

[project]
name = "kedro-brass"
readme = "README.md"
requires-python = ">=3.9"
version = "0.1.dev0"
license = {file = "LICENSE"}
description = "Write your Kedro pipelines and nodes in YAML"
dependencies = [
    "kedro>=0.19.12",
    "pydantic>=2.11.4",
    "pyyaml>=6.0",
    "structlog",
]
authors = [
    {name = "Juan Luis Cano Rodríguez", email = "hello@juanlu.space"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
source = "https://github.com/astrojuanlu/kedro-brass"
tracker = "https://github.com/astrojuanlu/kedro-brass/issues"
documentation = "https://kedro-brass.readthedocs.io"

[dependency-groups]
test = [
    "mypy",
    "hypothesis",
    "pytest",
    "pytest-cov",
]
doc = [
    "furo",
    "myst-parser",
    "sphinx>=5",
    "sphinx-copybutton",
]

[tool.ruff]
show-fixes = true

[tool.ruff.lint]
select = [
    "F",  # Pyflakes
    "E",  # Pycodestyle
    "W",  # Pycodestyle
    "D",  # Pydocstyle
    "UP",  # pyupgrade
    "I",  # isort
    "PL",  # Pylint
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["D", "PLR2004"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.mypy]
python_version = "3.9"
warn_redundant_casts = true
warn_unused_configs = true
pretty = true
show_error_codes = true
show_error_context = true
show_column_numbers = true

disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true

# More strict checks for library code
[[tool.mypy.overrides]]
module = "kedro_brass"
disallow_untyped_defs = true

# Ignore certain missing imports
# [[tool.mypy.overrides]]
# module = "thirdparty.*"
# ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
